[REFACTOR] Renamed Breadcrumb component as per naming standards
