+++ src/components/Breadcrumb/index.js
@@ -1,0 +1,3 @@
+import Breadcrumb from "./Breadcrumb.view"
+
+export default Breadcrumb;
